libs/nacos-client-python/pyproject.toml
[project]
name = "nacos-client-python"
version = "0.1.0"
description = "Python 工程，支持服务注册/发现到 Nacos 注册中心"
authors = [
    {name = "Your Name", email = "your@email.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "nacos-sdk-python>=2.0.0"
]

[tool.setuptools.packages.find]
where = ["src"] 